Se modifico la declaracion de fBuscaDireccionDocumento.

+++ Contpaqi/SistemasComerciales/Sdk/AdminpaqSDK.cs
@@ -141,7 +141,7 @@ namespace Contpaqi.SistemasComerciales.Sdk
         public static extern int fBuscaDireccionCteProv(string aCodCteProv, int aTipoDireccion);
 
         [DllImport("MGW_SDK.dll", EntryPoint = "fBuscaDireccionDocumento")]
-        public static extern int fBuscaDireccionDocumento(int aIdDocumento, int aTipoDireccion);
+        public static extern int fBuscaDireccionDocumento(int aIdDocumento, byte aTipoDireccion);
 
         [DllImport("MGW_SDK.dll", EntryPoint = "fBuscaDireccionEmpresa")]
         public static extern int fBuscaDireccionEmpresa();
